refactor: slim curtailment history pagination wiring
