--- client/src/protoFleet/features/energy/CurtailmentStatusDot.tsx
@@ -1,12 +1,0 @@
-import type { ReactElement } from "react";
-import clsx from "clsx";
-
-interface CurtailmentStatusDotProps {
-  className: string;
-}
-
-function CurtailmentStatusDot({ className }: CurtailmentStatusDotProps): ReactElement {
-  return <span className={clsx("inline-block h-2 w-2 shrink-0 rounded-full", className)} />;
-}
-
-export default CurtailmentStatusDot;
